pyproject.toml
[project]
name = "sports"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "benedict", email = "dobe.benedict@gmail.com"},
]
dependencies = ["numpy>=2.1.3", "pandas>=2.2.3", "seaborn>=0.13.2", "matplotlib>=3.9.2", "scikit-learn>=1.5.2", "dill>=0.3.9", "BeautifulSoup>=3.2.2", "tweepy>=4.14.0", "requests>=2.32.3", "Bs4>=0.0.2", "pyyaml>=6.0.2", "ensure>=1.0.4", "python-box>=7.2.0", "mlflow>=2.18.0"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
